+++ src/pages/ui/index.ts
@@ -1,4 +1,4 @@
-export { default as Display } from "./display.vue";
+export { default as ChatContent } from "./chat-content.vue";
 export { default as Submit } from "./submit.vue";
 export { default as Search } from "./search.vue";
 export { default as Save } from "./save.vue";
